+++ src/extension.ts
@@ -28,10 +28,14 @@ import { remoting } from './commands/remotingCommands';
 import { logging } from './commands/loggingCommands';
 import { MagitProcessLogEntry } from './models/magitProcessLogEntry';
 import { processView } from './commands/processCommands';
-import { resetting } from './commands/resettingCommands';
+import { resetting, resetMixed, resetHard } from './commands/resettingCommands';
 import { tagging } from './commands/taggingCommands';
 import { worktree } from './commands/worktreeCommands';
 import { diffing } from './commands/diffingCommands';
+import { ignoring } from './commands/ignoringCommands';
+import { running } from './commands/runningCommands';
+import { cherryPicking } from './commands/cherryPickingCommands';
+import { reverting } from './commands/revertingCommands';
 
 export const magitRepositories: Map<string, MagitRepository> = new Map<string, MagitRepository>();
 export const views: Map<string, DocumentView> = new Map<string, DocumentView>();
@@ -83,11 +87,17 @@ export function activate(context: ExtensionContext) {
   context.subscriptions.push(commands.registerTextEditorCommand('magit.merging', Command.primeRepo(merging)));
   context.subscriptions.push(commands.registerTextEditorCommand('magit.rebasing', Command.primeRepo(rebasing)));
   context.subscriptions.push(commands.registerTextEditorCommand('magit.resetting', Command.primeRepo(resetting)));
+  context.subscriptions.push(commands.registerTextEditorCommand('magit.reset-mixed', Command.primeRepo(resetMixed)));
+  context.subscriptions.push(commands.registerTextEditorCommand('magit.reset-hard', Command.primeRepo(resetHard)));
   context.subscriptions.push(commands.registerTextEditorCommand('magit.remoting', Command.primeRepo(remoting)));
   context.subscriptions.push(commands.registerTextEditorCommand('magit.logging', Command.primeRepo(logging)));
   context.subscriptions.push(commands.registerTextEditorCommand('magit.show-refs', Command.primeRepo(showRefs)));
   context.subscriptions.push(commands.registerTextEditorCommand('magit.diffing', Command.primeRepo(diffing)));
   context.subscriptions.push(commands.registerTextEditorCommand('magit.tagging', Command.primeRepo(tagging)));
+  context.subscriptions.push(commands.registerTextEditorCommand('magit.cherry-picking', Command.primeRepo(cherryPicking)));
+  context.subscriptions.push(commands.registerTextEditorCommand('magit.reverting', Command.primeRepo(reverting)));
+  context.subscriptions.push(commands.registerTextEditorCommand('magit.ignoring', Command.primeRepo(ignoring)));
+  context.subscriptions.push(commands.registerTextEditorCommand('magit.running', Command.primeRepo(running)));
   context.subscriptions.push(commands.registerTextEditorCommand('magit.worktree', Command.primeRepo(worktree)));
   context.subscriptions.push(commands.registerTextEditorCommand('magit.process-log', Command.primeRepo(processView, false)));