+++ src/models/magitCherryPickingState.ts
@@ -1,0 +1,7 @@
+import { Commit } from '../typings/git';
+
+export interface MagitCherryPickingState {
+  originalHead: Commit;
+  currentCommit: Commit;
+  upcomingCommits: Commit[];
+}