+++ src/views/dispatchView.ts
@@ -14,11 +14,12 @@ export class DispatchView extends DocumentView {
     super(uri);
 
     this.addSubview(new TextView(`Popup and dwim commands
-    d Diffing           b Branching        c Committing
-    f Fetching          F Pulling          l Logging 
-    m Merging           M Remoting         t Tagging
-    P Pushing           r Rebasing         X Resetting
-    y Show Refs         z Stashing         % Worktree`));
+    A Cherry-picking    b Branching         c Committing
+    d Diffing           f Fetching          F Pulling
+    i Ignoring          l Logging           m Merging
+    M Remoting          P Pushing           r Rebasing
+    t Tagging           V Reverting         X Resetting
+    y Show Refs         z Stashing          ! Running           % Worktree`));
   }
 
   public update(state: MagitState): void { }